Revert "Move ShardingSphereDatabasesFactory to mode-core module (#35120)"
This reverts commit 9fb2035ec21e98e627d26d22c1cb04ae60d86333.
